+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### changed
+- Requires [`httpx`](https://www.python-httpx.org)==0.20.\*
 
 ## [0.13.0] - 2021-08-19
 ### Changed